+++ src/services/cart/InitCartOperator.ts
@@ -1,0 +1,8 @@
+import { CartOperator } from './CartOperator'
+import { CartOperatorEnum } from './CartOperatorEnum'
+
+export class InitCartOperator extends CartOperator {
+  async operation() {
+    this.syncCartProducts(CartOperatorEnum.INIT)
+  }
+}
