+++ src/services/cart/RemoveCartProductOperator.ts
@@ -1,0 +1,11 @@
+import { CartOperator } from './CartOperator'
+import { CartOperatorEnum } from './CartOperatorEnum'
+
+export class RemoveCartProductOperator extends CartOperator {
+  async operation(productIds: string[]) {
+    const cachedCartProducts = this.getLocalCartProducts()
+    const newCartProduct = cachedCartProducts.filter(cartProduct => !productIds.includes(cartProduct.productId))
+    localStorage.setItem('kolable.cart._products', JSON.stringify(newCartProduct))
+    this.syncCartProducts(CartOperatorEnum.REMOVE_CART_PRODUCTS)
+  }
+}
